test_helpers.go file for node package
